+++ vignettes/_figures/make_README_plots.R
@@ -22,12 +22,8 @@ save_it = function(filename) {
 library(mcp)
 options(mc.cores = 3)  # Run in parallel
 
-model_demo = list(
-  response ~ 1,  # plateau (int_1)
-  ~ 0 + time,  # joined slope (time_2) at cp_1
-  ~ 1 + time  # disjoined slope (int_1, time_2) at cp_2
-)
-fit_demo = mcp(model_demo, data = ex_demo, adapt = 3000)  # dataset included in mcp
+ex_demo = mcp_example("demo")
+fit_demo = mcp(model_demo$model, data = ex_demo$data, adapt = 3000)  # dataset included in mcp
 theme_it(plot(fit_demo), "")
 save_it("ex_demo.png")
 
@@ -40,7 +36,7 @@ model_null = list(
   response ~ 1 + time,
   ~ 1 + time
 )
-fit_null = mcp(model_null, ex_demo, adapt = 3000)
+fit_null = mcp(model_null, ex_demo$data, adapt = 3000)
 
 # Compare loos:
 fit_demo$loo = loo(fit_demo)
@@ -51,11 +47,8 @@ loo::loo_compare(fit_demo$loo, fit_null$loo)
 ################
 # Two plateaus #
 ################
-model_plateaus = list(
-  y ~ 1,  # plateau (int_1)
-  ~ 1  # plateau (int_2)
-)
-fit_plateaus = mcp(model_plateaus, ex_plateaus, par_x = "x", adapt = 3000)
+ex_intercepts = mcp_demo("intercepts")
+fit_plateaus = mcp(ex_intercepts$model, ex_intercepts$data, par_x = "x", adapt = 3000)
 theme_it(plot(fit_plateaus, lines = 25), "Two plateaus")
 save_it("ex_plateaus.png")
 
@@ -64,12 +57,8 @@ save_it("ex_plateaus.png")
 ########################
 # VARYING SLOPE CHANGE #
 ########################
-
-model_varying = list(
-  y ~ 1 + x,  # intercept + slope
-  1 + (1|id) ~ 0 + x + sigma(1)  # joined slope, varying by id
-)
-fit_varying = mcp(model_varying, ex_varying, adapt = 3000)
+ex_varying = mcp_example("varying")
+fit_varying = mcp(ex_varying$model, ex_varying$data, adapt = 3000)
 theme_it(plot(fit_varying, facet_by = "id"), "Varying slope change")
 save_it("ex_varying.png")
 
@@ -77,12 +66,8 @@ save_it("ex_varying.png")
 ############
 # BINOMIAL #
 ############
-model_binomial = list(
-  y | trials(N) ~ 1,  # constant rate
-  ~ 0 + x,  # joined changing rate
-  ~ 1 + x  # disjoined changing rate
-)
-fit_binomial = mcp(model_binomial, ex_binomial, family = binomial(), adapt = 5000)
+ex_binomial = mcp_example("binomial")
+fit_binomial = mcp(ex_binomial$model, ex_binomial$data, family = binomial(), adapt = 5000)
 theme_it(plot(fit_binomial, q_fit = TRUE), "Binomial")
 save_it("ex_binomial.png")
 
@@ -90,11 +75,7 @@ save_it("ex_binomial.png")
 ##########################
 # FIXED, RELATIVE, PRIOR #
 ##########################
-model_rel = list(
-  y ~ 1 + x,
-  ~ rel(1) + rel(x),
-  rel(1) ~ 0 + x
-)
+ex_rel = mcp_example("rel_prior")
 prior_rel = list(
   int_1 = 10,  # fixed value
   x_3 = "x_1",  # shared slope in segment 1 and 3
@@ -102,7 +83,7 @@ prior_rel = list(
   cp_1 = "dunif(20, 50)"  # has to occur in this interval
 )
 
-fit_rel = mcp(model_rel, ex_rel_prior, prior_rel, iter = 10000)
+fit_rel = mcp(ex_rel$model, ex_rel$data, prior_rel, iter = 10000)
 theme_it(plot(fit_rel, cp_dens = FALSE), "rel() and prior")
 save_it("ex_fix_rel.png")
 
@@ -110,10 +91,7 @@ save_it("ex_fix_rel.png")
 #############
 # QUADRATIC #
 #############
-model_quadratic = list(
-  y ~ 1,
-  1 ~ 0 + x + I(x^2)
-)
+ex_quadratic = mcp_example("quadratic")
 fit_quadratic = mcp(model_quadratic, ex_quadratic, adapt = 3000)
 theme_it(plot(fit_quadratic), "Quadratic and other exponentiations")
 save_it("ex_quadratic.png")
@@ -123,12 +101,8 @@ save_it("ex_quadratic.png")
 #################
 # TRIGONOMETRIC #
 #################
-# Define the model
-model_trig = list(
-  y ~ 1 + sin(x),
-  ~ 0 + cos(x) + x
-)
-fit_trig = mcp(model_trig, ex_trig, adapt = 3000)
+ex_trig = mcp_example("trigonometric")
+fit_trig = mcp(ex_trig$model, ex_trig$data, adapt = 3000)
 theme_it(plot(fit_trig), "Trigonometric for periodic trends")
 save_it("ex_trig.png")
 
@@ -137,13 +111,8 @@ save_it("ex_trig.png")
 ############
 # VARIANCE #
 ############
-model_variance = list(
-  y ~ 1,
-  ~ 0 + sigma(1 + x),
-  ~ 0 + x
-)
-
-fit_variance = mcp(model_variance, ex_variance, iter = 10000, adapt = 10000)
+ex_variance = mcp_example("variance")
+fit_variance = mcp(ex_variance$model, ex_variance$data, iter = 10000, adapt = 10000)
 theme_it(plot(fit_variance, q_predict = TRUE), "Variance and prediction intervals")
 save_it("ex_variance.png")
 
@@ -152,12 +121,7 @@ save_it("ex_variance.png")
 #########
 # AR(N) #
 #########
-model_ar = list(
-  price ~ 1 + ar(2),
-  ~ 0 + time + ar(1)
-)
-
-fit_ar = mcp(model_ar, ex_ar, adapt = 3000)
-plot_ar = plot(fit_ar)
-theme_it(plot_ar, "Time series with autoregressive residuals")
+ex_ar = mcp_example("ar")
+fit_ar = mcp(ex_ar$model, ex_ar$data, adapt = 3000)
+theme_it(plot(fit_ar), "Time series with autoregressive residuals")
 save_it("ex_ar.png")